+++ src/EntityManager.php
@@ -22,6 +22,7 @@ use GraphAware\Neo4j\OGM\Metadata\GraphEntityMetadata;
 use GraphAware\Neo4j\OGM\Metadata\QueryResultMapper;
 use GraphAware\Neo4j\OGM\Metadata\RelationshipEntityMetadata;
 use GraphAware\Neo4j\OGM\Repository\BaseRepository;
+use GraphAware\Neo4j\OGM\Repository\ObjectHydration;
 use GraphAware\Neo4j\OGM\Util\ClassUtils;
 
 class EntityManager implements ObjectManager
@@ -224,6 +225,12 @@ class EntityManager implements ObjectManager
     public function getUnitOfWork()
     {
         return $this->uow;
+    }
+
+    public function getHydrator($mode='')
+    {
+        // TODO imprive this. Maybe need different type of hydrators.
+        return new ObjectHydration($this);
     }
 
     /**
